+++ src/components/HeroSlide/HeroSlide.jsx
