+++ src/components/SearchIcon/SmallHighlightSection.jsx
@@ -1,0 +1,6 @@
+
+export const SmallHighlightSection = () => {
+  return (
+    <div>SmallHighlightSection</div>
+  )
+}
